Added title-fall-state.c which makes the word fall logo falling letters into a state of its own. title-state.c now does not move at all an will be the default 'static' state for the hover colors to fall back on.
When the user switches back to the main screen from in the game, I am now
going to have to call state-title-fall if I want to see the letters fall
down.

I keep adding to the number of states, which I am generally trying to keep
low because I am not using oop for the states, but I think making the logo
transition into a state was a good itea because it makes much more sense.