Main feature implemented: Highlighting on title screen
Added Files:
title-highlight-options-state.c
title-highlight-play-state.c

Highlighting is now working for mouse overs on the play and options buttons.
I also added logic to title-state title-fall-state title-highlight-play-
state and title-highlight-options-state that allows the user to highlight
or click a button while the logo is still falling down on the screen. For
the logo to be seen falling from the beginning STATE_TITLE_FALL still must
be called.

This commit may be the end of the collision branch, which I will merge
into master, or I may keep the brach open, finish the options menu, and
then make a new branch for the play state.