Added new check button
Changed the color of the check sumbit button from green to hellow so
that the check does not blend in with the grass. There are two versions
of the check: one with a black outline and one with a yellow outline. I
will have to test both to see which one looks better.