Code Cleanup
Huge code cleanup throughout the program. The biggest change is the
move to an array that holds the title logo. This array is kind of wonky
in that I couldn't figure out how to define it with an extern, so
there is an awkward definition in init(). This array is also global,
but I think that this is a huge improvement over my old code.

I really need to reconsider my state system, because I am not passing
any values into any of the funtions, which is forcing me to use
global variables for anything that is used by more than one function.