Code Cleanup -- Fixed some global variables
I finally fixed the issues with the extern title array in main.h. There
are now only a few global values that are all declared in main.h and
defined at the top of main.c. I don't see any good way around not
using these globals but as long as I am careful, they shouldn't be a
problem.