Transitions Finish and Code Cleanup
game-transition-state.c is finished. All of the transitions for the
title, fonts, container, and pause button have been implemented.

I also created two header files game-state.h and title-state.h which
contain extern definitions of variables that are shared between files
in each of their respective states. I believe that this removes all but
two of my global variables in main.h. This makes the project scope a
little smaller, which shoudld help out with overall organization of the
project.

The next step if I have time would be to do to the surfaces what I did
to the state variables, and move them into their respective header
files instead of having them all be global. I'm not going to worry
about this until I have time.