Submit Button Rect and Bug Fixes
Collision detection for the submit button was added. It currently just
uses a print statement.

I changed the behavior of the fall state so that the user is forced to
wait until the Word Fall logo has fallen completely before they are
allowed to press play. This is mainly so that my animation after
clicking play makes sense. During user testing, if the user thinks
that the game is lagging because the play button doesn't respond for
what seems like a few seconds,(but is actually just until the animation
finishes) I will change it back, but right now I can't imagine anyone
wanting to click the play button that badly in the beginning. So
hopefully this won't be a problem.