Letter Dragging Implemented
Dragging letters has been implemented. The framework for having all of
the random letters falling has been implemented, I just need to set it
up. There are a few dirty fixes I made to implement dragging with just
one letter that will have to be cleaned up for all of the letters.

Another thing I am going to have to implement is a snap feature for the
letters, so that when the user gets close to one of the submit
continers, the letter will automatically go into it instead of having
the user meticulously drop it into the box.