Added One Falling Letter
Currently, a random letter is chosen and it falls, but there will need
to be some major revisions to get all of the random letters to fall.
In order to have multiple of the same letter on the screen at once, I
am going to have to add the the letters array and make it
multi-dimensional.