Added Multidimensional Array For Letters
The letters falling have been completely changed. Now there are 4
surfaces. A random spot is chosen, and then a random letter so that
a random letter ends up coming out of one of the four spots.

The biggest challenge with this is going to be keeping track of all of
the letters that are falling, so that I can make them continue to fall
because currently, I am just incrementing the random array that was
chosen, but if I overwrite that, the letter stops falling.

The other big things that will need to be implemented are the snap
into contianer feature as well as the sumbit button lookup.