All Random Falling Letters Implemented
All of the random falling letters have been implemented. The letters
are currently only going one pixel per cycle, but that is turning out
to be pretty fast. If I switch to floating point numbers, using
decimals might slow it down, but it that doesn't work, I am going to
need to increment the numbers only after a certain number of seconds
with a timer.

Still need to implement snap feature as well as submit button.