Snap Feature Implemented
Currently, after the letter snaps into place, when you let go of it, it
disappears because I am drawing the grass in front of the letters.
What I will have to do is create a whole new array of surfaces that
just holds the letters that are in the container. This way I can just
constantly draw any letters that are in that surface underneath the
contianer.

I will also have to take into consideration people dragging letters
out from the container as well as moving the order of the letters in
the container, so I will have to modify my array to reflect that.