Words can now be dropped into the containers
There are still a bunch of bug fixes that I need to make, but
letters can now be dropped into the container. The biggest thing
is my "snap" feature which only detects from the top left corner, so
if a user has 3 corners in, but not the top left corner, then the
collision won't register. This will likely have to be changed.