Framework For Submit Button Implemented
Hitting the submit button will printf the ascii value of the
containers with all spaces in the containers corresponsing to ascii
spaces.

Next I am going to need to find a way to get that information to the
python script so that I can truncate any spaces on the end, or reject
any spaces in the middle, and then check the dictionary.