Dictionary Checks Implemented
Now when the user clicks on the submit button, the word will be
checked against the dictionary and if it is a word, the function will
return 1 or if there isn't a word the funciton will return 0.

Right now, there is just a print statement indicating if there is a
word or not, but eventually I will need to have scores pop up on the
screen for each correct word.

Before I send a word to be checked against the dictionary, I will
need to sanitize the input first. Right now I have it where any
blank tiles correspond to blank spaces in the ascii string. All blank
spaces must be removed before passing the word to the function. I will
also need to check if there is a space in between letters and if there
is, reject the word immediately because spaces in the middle are not
allowed.