Added surfaces and variables for score
I added all of the fonts and variables necessary for keeping score. Now
I am going to have to keep track of the score by keeping an array of
the points that corresponds to the values in score value. I will need to
calculate the score by messing with the ascii value to find which letter
I am dealing with.