Added score/no-score transitions and total score
The total score is displayed at the top next to the pause button, but
it needs a better spot because the clouds are in the way.

The score transitions will pop up and then fade away with the amount
of points you earned from the submit. If the word is invalid there will
be an x that pops up.

Right now I really need to figure out how to make it spit out random
words so that it is even remotely fun to play, or at least raise the
probablility of finding common letters. I seriously waited over a
minute just to find a vowel that I needed. Unfortunately, the increased
probablility numbers are going to be hard, and the random words
might be even harder.

Other than changing the way that the words fall down, and adding
music, the game is basically done. There will need to be some
finishing touches like the pause and options menu, as well as a
highscore. I am almost done with this game. It isn't great, but it is
still the best thing I have ever made, and I does have really good
transitions. I just want to get to a point where I can call the game
completed, even if it isn't actually that fun.