Removed Magic Numbers for some fonts
I had the positions of the fonts hard coded. I changed this so that
they would be calculated relative to other objects so that if I change
the fonts to upper case, I won't have formatting issues. I am using
divisions that are being comuted 60 times a second, which may be a
performance bottleneck, but it might be optimized by the compiler so
that it is a non-issue. This release may have introduced some bugs
surrounding the placement of the fonts.

There is still a lot to do. The biggest challenges are going to  be the
pause screen as well as the sound. If I create a new state for the
pause button(which I should) it would require me to retain the (x,y)
values of the letters and other objects on the game-state page because
they would be present in the pause state. I am tempted to just put a
huge if statement around the whole thing that says if(!paused) but that
would be bad practice. I need to just suck it up and create a new state.
Also, final code cleanup is going to be a bitch. I just want to be done,
but the codebase is getting messier by the minute.