Pause Button Implemented.
The pause button has been implemented. When the user clicks the pause
button in the game a popup will say that the game is paused, and
everything in the background will stop moving. The use can then click
the resume button, or the pause button again to continue where they
left off.

I also changed the position of the clouds. All of the clouds are
slightly lower now to accommodate for the score and pause button. The
variable for the clouds is the global struct cloud. I know it sholdn't
be global, but it was so much easier and made sense because half of the
states in the code contained the clouds. Having inheriance and
protected vaariables like in C++ would be extremenly valuable here.

TODO:
1. sounds
2. save states(scores,sound preference)
3. display highscore
4. Code cleanup

I might also need to make some visual adjustments like moving the word
fall title down to compensate for the lowered clouds, as well as fixing
the borders around the letters.