Sound and Bug Fixes
Background music and sound effects have been added. I also squashed a
bug that was causing letters to appear from behind the cloud. Also,
now the user can't click on the letter containers in the cloud and
drag letters down. They now must wait until that letter was randomly
chosen as intended.

The only other feature implementations left are save states(score,
and sound preferences) as well as a button to return to the main
menu from the pause state.

Current Bugs.
Keeping the letter in the screen still sucks. It may have something to
do with me not checking boundaries while the left mouse button is
held up. Also when they release the LMB, I will need to put the letter
back into the screen to ensure that the user didn't drag it into
oblivion.

My next big challenge is going to be letting the user drag the words
from outside the container. This is just a no brainer and I have
decided that it needs to be implemented. What I am going to have to do
is keep a separate array for letters in the container(I may already
have this) and then let the user take the letter out of the container
and put it in another spot, or throw it up into the screen and have
it fall down with all of the other letters. Hopefully this shouldn't
be that difficult, because it is going to improve the gameplay
immensely.