Added return to main menu from pause screen
A button to return to the main menu has been added, but it has
introduced a whole host of new, unknown bugs. The largest is that I
don't reset the value of the variables on state change, which means
that if the user goes to the main menu, and then presses play, he or
she is taken back to the current game. I need to reset all of these
variables. The only other main feature implementations are the
highscore saves/display of highscore and some minor settings in
the options menu. I really need to finish feature implementation
because I have a feeling that code cleanup and bug fixes are going
to take an inordinate amount of time.

Bugs:
1. Game doesn't reset variables on state change. This is where the
constructors and destructors of objects would have been helpful.
2. Game now crashes randomly. There aren't any steps to reprodcue, but
it does seem to happen frequently while dragging letters around.