Added Lua Dictionary Checks
Lua is now used to check the dictionary to see if a word exists. This
means that the large dict buffer that I had running for the length of
the program is now gone and has been replaced by a call to a lua
script.

The lua script reads the list of words every time it is called
which I am sure is horribly inefficient. I might be able to speed this
up if I were able to pass the lua-read data back into C so that I
can keep it in memory. This would require a whole lot of work, and is
most likely not going to happen, because it looks like I don't need the
optimization. There is a slight delay after hitting submit because of
the call to Lua, and ideally it would be faster, but it is fast enough.

-- NOTE --
Word Fall now requires linking with -llua. I will need to figure out
how to distribute this with my application.