Changed compiler to Clang
The /usr/bin/cc files links to clang, while the /usr/bin/gcc links to
the llvm-gcc compiler. I have switched to clang completely for OSX.
The errors are nicer, and hopefully, it will produce more stable
code. Clang did complain that theere wasn't a newline at the end
of all of the files, so I added a newline. I think this is a rather
strange warning, but I'll go with it.

I also changed the strncpy in the sanitize function to use a strcpy,
hopefully this doesn't create any adverse affects.