Initial Commit
This branch contains all of the files that were needed to create
the windows installer, which means that this is a fully working
copy of Word Fall on windows. This branch also contains the
os x background image for the dmg. The script file for the Inno
Setup installer is also included.