Added Hash Table To Check Dictionary Values
Warning:
There is currently a massive memory leak in word fall that is
causing it to eat around 3MB of ram per second, and the leak
does not stop. The program will easily consume gigabytes of ram
in a matter of minutes. The cause of this bug is still being
investigated.