Fixed Memory Leak
Fonts that were being rendered were not being freed. So a pointer to a
new surface was being malloc and a pointer would point to the memory.
Next game loop, the pointer was changed without freeing the
font so the memory was lost. This has been fixed by freeing the pointer
immediately after it is done being used.